+++ src/main/java/com/knu/KnowcKKnowcK/controller/articleSummary/OpinionController.java
@@ -4,14 +4,13 @@ import com.knu.KnowcKKnowcK.apiResponse.ApiResponseDto;
 import com.knu.KnowcKKnowcK.apiResponse.SuccessCode;
 import com.knu.KnowcKKnowcK.dto.requestdto.OpinionRequestDto;
 import com.knu.KnowcKKnowcK.dto.responsedto.OpinionResponseDto;
-import com.knu.KnowcKKnowcK.service.articleOpinion.SaveOpinionService;
+import com.knu.KnowcKKnowcK.service.opinion.OpinionService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -24,7 +23,7 @@ import org.springframework.web.bind.annotation.RestController;
 @Tag(name = "견해 작성 관련 API", description = "견해 관련 API")
 public class OpinionController {
 
-    private final SaveOpinionService saveOpinionService;
+    private final OpinionService opinionService;
 
     @PostMapping("/submit")
     @Operation(summary = "요약 피드백 받기", description = "작성한 요약에 대한 AI  피드백에 제공됨")
@@ -34,7 +33,7 @@ public class OpinionController {
     })
     ApiResponseDto<OpinionResponseDto> getOpinionFeedback(Authentication authentication, @RequestBody @Valid OpinionRequestDto dto){
 
-        OpinionResponseDto opinionFeedback = saveOpinionService.getOpinionFeedback(dto, authentication.getName());
+        OpinionResponseDto opinionFeedback = opinionService.getOpinionFeedback(dto, authentication.getName());
         return ApiResponseDto.success(SuccessCode.OK, opinionFeedback);
     }
