--- src/main/java/com/knu/KnowcKKnowcK/service/articleOpinion/SaveOpinionService.java
@@ -1,10 +1,0 @@
-package com.knu.KnowcKKnowcK.service.articleOpinion;
-
-import com.knu.KnowcKKnowcK.dto.requestdto.OpinionRequestDto;
-import com.knu.KnowcKKnowcK.dto.responsedto.OpinionResponseDto;
-
-public interface SaveOpinionService {
-
-    OpinionResponseDto getOpinionFeedback(OpinionRequestDto dto, String writer);
-
-}
