--- src/main/java/com/knu/KnowcKKnowcK/service/articleSummary/LoadArticlesService.java
@@ -1,16 +1,0 @@
-package com.knu.KnowcKKnowcK.service.articleSummary;
-
-import com.knu.KnowcKKnowcK.domain.Article;
-import com.knu.KnowcKKnowcK.dto.responsedto.article.ArticleListResponseDto;
-import com.knu.KnowcKKnowcK.enums.Category;
-import org.springframework.data.domain.Page;
-
-import java.util.Optional;
-
-public interface LoadArticlesService {
-
-    Page<ArticleListResponseDto> loadArticles(Category category, int page, String memberEmail);
-
-    Article loadArticleById(Long id);
-
-}
