--- src/main/java/com/knu/KnowcKKnowcK/service/articleSummary/LoadSummaryService.java
@@ -1,8 +1,0 @@
-package com.knu.KnowcKKnowcK.service.articleSummary;
-
-import com.knu.KnowcKKnowcK.dto.responsedto.SummaryHistoryResponseDto;
-
-public interface LoadSummaryService {
-
-    SummaryHistoryResponseDto loadSummaryHistory(String userEmail, long articleId);
-}
