--- src/main/java/com/knu/KnowcKKnowcK/service/articleSummary/SaveSummaryService.java
@@ -1,11 +1,0 @@
-package com.knu.KnowcKKnowcK.service.articleSummary;
-
-import com.knu.KnowcKKnowcK.dto.requestdto.SummaryRequestDto;
-import com.knu.KnowcKKnowcK.dto.responsedto.SummaryResponseDto;
-
-public interface SaveSummaryService {
-
-  SummaryResponseDto saveSummary(SummaryRequestDto dto, String writer);
-
-  SummaryResponseDto getSummaryFeedback(SummaryRequestDto dto, String writer);
-}
